Add interactive moves history
