+++ src/App.js
@@ -1,80 +1,50 @@
 import { useState } from "react";
-import Square from "./components/Square";
-import { calculateWinner } from "./helpers/helpers";
+import Board from "./components/Board";
 import "./styles.css";
 
-export default function Board() {
-  const [squares, setSquares] = useState(Array(9).fill(null));
-  const [xIsNext, setXIsNext] = useState(true);
+export default function Game() {
+  const [history, setHistory] = useState([Array(9).fill(null)]);
+  const [currentMove, setCurrentMove] = useState(0);
+  const xIsNext = currentMove % 2 === 0;
+  const currentSquares = history[currentMove];
 
-  const players = {
-    x: "X",
-    o: "O",
-  };
-
-  function handleClick(i) {
-    if (squares[i] || calculateWinner(squares)) return;
-
-    const nextSquares = squares.slice();
-
-    xIsNext ? (nextSquares[i] = players.x) : (nextSquares[i] = players.o);
-
-    setSquares(nextSquares);
-    setXIsNext(!xIsNext);
+  function handlePlay(nextSquares) {
+    const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
+    setHistory(nextHistory);
+    setCurrentMove(nextHistory.length - 1);
   }
 
-  const winner = calculateWinner(squares);
-  let status;
+  function jumpTo(nextMove) {
+    setCurrentMove(nextMove);
+  }
 
-  winner
-    ? (status = `Winner: ${winner}`)
-    : (status = `Next Player: ${xIsNext ? players.x : players.o}`);
+  const moves = history.map((squares, move) => {
+    let description;
 
+    if (move > 0) {
+      description = `Go to move # ${move}`;
+    } else {
+      description = `Go to start`;
+    }
+
+    return (
+      <li key={move}>
+        <button onClick={() => jumpTo(move)}>{description}</button>
+      </li>
+    );
+  });
   return (
-    <>
-      <h4>{status}</h4>
-      <div className='board-row'>
-        <Square
-          value={squares[0]}
-          onSquareClick={() => handleClick(0)}
-        />
-        <Square
-          value={squares[1]}
-          onSquareClick={() => handleClick(1)}
-        />
-        <Square
-          value={squares[2]}
-          onSquareClick={() => handleClick(2)}
+    <div className='game'>
+      <div className='game-board'>
+        <Board
+          xIsNext={xIsNext}
+          squares={currentSquares}
+          onPlay={handlePlay}
         />
       </div>
-      <div className='board-row'>
-        <Square
-          value={squares[3]}
-          onSquareClick={() => handleClick(3)}
-        />
-        <Square
-          value={squares[4]}
-          onSquareClick={() => handleClick(4)}
-        />
-        <Square
-          value={squares[5]}
-          onSquareClick={() => handleClick(5)}
-        />
+      <div className='game-info'>
+        <ol>{moves}</ol>
       </div>
-      <div className='board-row'>
-        <Square
-          value={squares[6]}
-          onSquareClick={() => handleClick(6)}
-        />
-        <Square
-          value={squares[7]}
-          onSquareClick={() => handleClick(7)}
-        />
-        <Square
-          value={squares[8]}
-          onSquareClick={() => handleClick(8)}
-        />
-      </div>
-    </>
+    </div>
   );
 }
